Replace outdating sharing content test
